requirements.txt
Django>=5.2.3,<6.0
djangorestframework>=3.14.0,<4.0
Pillow>=10.0.0,<11.0
PyPDF2>=3.0.0,<4.0
PyMuPDF>=1.22.0,<2.0
python-magic>=0.4.27,<0.5
numpy>=1.24.0,<2.0
